requirements.txt
uvicorn==0.16.0
fastapi==0.70.1
python-jose==3.3.0
fastapi-utils==0.2.1
gunicorn==20.0.4
PyJWT==1.7.1
python-multipart==0.0.5